requirements.txt
google-generativeai
flask
flask-cors